Refactor BLS change pool
